Fast-API-Backend/genres_styles/requirements.txt
fastapi
uvicorn
torch
torchvision
timm
pillow
numpy
gdown
vercel-fastapi
